NOJIRA:  Split out harness and worker for cleaner reuse.
