+++ tests/js/lib/index.js
@@ -1,2 +1,3 @@
 "use strict";
+require("./worker");
 require("./harness");
